calendar-backend/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
pydantic
